replaced tikToken usage with text-embedding-ada-002
for representing db schema in a vector and added a new column to the vector DB text_schema.

+++ src/utils/types.py
@@ -1,1 +1,1 @@
-embedding = list[int]
+embedding = list[float]
